adds marked js package  to parse the rich text  in grid A component. rename theme file to styles.
