+++ package.json
@@ -10,6 +10,7 @@
     "@mui/styles": "^5.7.0",
     "cra-template": "1.1.3",
     "intersection-observer": "^0.12.0",
+    "marked": "^4.0.16",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-router-dom": "^6.3.0",
