+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 import { ThemeProvider } from '@mui/material/styles';
-import theme from './theme/theme';
+import theme from './styles/theme';
 import './index.css';
 import App from './App';
